MAGE-986 Move new config options to discrete InstantSearch config helper
