+++ Helper/ConfigHelper.php
@@ -47,9 +47,6 @@ class ConfigHelper
     public const INFINITE_SCROLL_ENABLE = 'algoliasearch_instant/instant_options/infinite_scroll_enable';
     public const HIDE_PAGINATION = 'algoliasearch_instant/instant_options/hide_pagination';
 
-    public const IS_INSTANT_REDIRECT_ENABLED = 'algoliasearch_instant/instant_redirects/enable';
-    public const INSTANT_REDIRECT_OPTIONS = 'algoliasearch_instant/instant_redirects/options';
-
     public const IS_POPUP_ENABLED = 'algoliasearch_autocomplete/autocomplete/is_popup_enabled';
     public const NB_OF_PRODUCTS_SUGGESTIONS = 'algoliasearch_autocomplete/autocomplete/nb_of_products_suggestions';
     public const NB_OF_CATEGORIES_SUGGESTIONS = 'algoliasearch_autocomplete/autocomplete/nb_of_categories_suggestions';
@@ -1061,21 +1058,6 @@ class ConfigHelper
             ScopeInterface::SCOPE_STORE,
             $storeId
         );
-    }
-
-    public function isInstantRedirectEnabled(?int $storeId = null): bool
-    {
-        return $this->configInterface->isSetFlag(self::IS_INSTANT_REDIRECT_ENABLED, ScopeInterface::SCOPE_STORE, $storeId);
-    }
-
-    public function getInstantRedirectOptions(?int $storeId = null): array
-    {
-        $value = $this->configInterface->getValue(
-            self::INSTANT_REDIRECT_OPTIONS,
-            ScopeInterface::SCOPE_STORE,
-            $storeId
-        );
-        return empty($value) ? [] : explode(',', $value);
     }
 
     /**
